Update meta.json
The original Chainspace Discord had a custom URL (/chainspace) but lost it when our Boosts went below a threshold, and some hacker picked up the URL and now has an active wallet drainer in that Discord. Please accept this commit ASAP so that we can avoid more people getting drained.